more malloc

+++ 0x0C-more_malloc_free/main.h
@@ -15,7 +15,7 @@ void *_calloc(unsigned int nmemb, unsigned int size);
 int *array_range(int min, int max);
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size);
 int _atoi(const char *s);
-void print_int(int n);
+void print_int(unsigned long int n);
